refactor(cli-upload): fold the dimension read into upload.js
Drops the `image-size.js` adapter, the `fixtures.js` module and the
`image-size` unit specs. Those existed to hold and prove a hand-rolled
PNG/JPEG parser; with `probe-image-size` doing the parsing, they covered
upstream's segment walking rather than anything this package owns.

What remains of the adapter is a bounded header read and a format gate, both
short enough to live beside their only caller. The four image fixtures the
end-to-end specs use move inline, and a spec covers the format gate directly:
a GIF named `.png` clears the extension filter and parses fine, so only the
gate keeps it out.

The change to `percy upload` is now the dependency swap plus the skip path.
